housekeeping in "bytes" variables types. resolve a int boundary bug.

+++ cookies.h
@@ -10,4 +10,4 @@ int listcookies(cookie_t *cookies);
 char* getcookie(cookie_t *cookies, char *name);
 char* generate_cookies_string(cookie_t *cookies, char *buf, unsigned int buflen);
 int freecookies(cookie_t *cookies);
-int parsehttpheadersforgettingcookies(cookie_t *cookies, const char *httpheaders, unsigned int bytes);
+int parsehttpheadersforgettingcookies(cookie_t *cookies, const char *httpheaders, ssize_t bytes);
